+++ spacegraphcats/conf/twofoo-short.yaml
@@ -1,12 +1,12 @@
 catlas_base: twofoo-short
 input_sequences:
-- spacegraphcats/search/test-data/63.short.fa.gz
-- spacegraphcats/search/test-data/twofoo-short-reads.fa.gz
+- tests/test-data/63.short.fa.gz
+- tests/test-data/twofoo-short-reads.fa.gz
 ksize: 31
 radius: 1
 search:
-- spacegraphcats/search/test-data/2.short.fa.gz
-- spacegraphcats/search/test-data/47.short.fa.gz
-- spacegraphcats/search/test-data/63.short.fa.gz
+- tests/test-data/2.short.fa.gz
+- tests/test-data/47.short.fa.gz
+- tests/test-data/63.short.fa.gz
 searchquick:
-- spacegraphcats/search/test-data/63.short.fa.gz
+- tests/test-data/63.short.fa.gz
