+++ tests/pytest_utils.py
@@ -34,7 +34,7 @@ class TempDirectory(object):
 def relative_file(filename):
     "Return the filename relative to the top level directory of this repo."
     thisdir = os.path.dirname(__file__)
-    pkgdir = os.path.join(thisdir, "../..")
+    pkgdir = os.path.join(thisdir, "..")
     newpath = os.path.join(pkgdir, filename)
     return os.path.abspath(newpath)
