+++ tests/test-data/47.short.fa.gz
