+++ tests/test-data/catlas.dory_k21_r1/contigs.fa.gz
