+++ tests/test-data/catlas.dory_k21_r1_search_oh0/dory-head.fa.cdbg_ids.txt.gz
