+++ tests/test-data/twofoo-short-reads.fa.gz
